+++ entry/views.py
@@ -401,8 +401,8 @@ def refer(request, entry_id):
 
 
 @airone_profile
-@http_get
-def export(request, entity_id):
+@http_post([])
+def export(request, entity_id, recv_data):
     user = User.objects.get(id=request.user.id)
 
     job_params = {
@@ -413,7 +413,7 @@ def export(request, entity_id):
     if not Entity.objects.filter(id=entity_id).exists():
         return HttpResponse('Failed to get entity of specified id', status=400)
 
-    if 'format' in request.GET and request.GET.get('format') == 'CSV':
+    if 'format' in recv_data and recv_data['format'] == 'CSV':
         job_params['export_format'] = 'csv'
 
     # check whether same job is sent